app/requirements.txt
fastapi~=0.115.7
uvicorn~=0.34.0
numpy~=2.0.2
opencv-python~=4.11.0.86
opencv-python-headless~=4.11.0.86
pillow~=11.1.0
python-multipart~=0.0.20